iss: Changed cosimulation project structure to workspaces to split broker from tui
