+++ vadl-cosim/.gitignore
@@ -20,3 +20,5 @@ target/
 /target
 
 .idea/
+
+!src/bin
