+++ vadl-cosim/cosim-lib/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "cosim-lib"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = "1.0.98"
+
+clap = { version = "4.5.41", features = ["derive"] }
+figment = { version = "0.10.19", features = ["toml"] }
+
+libc = "0.2.174"
+
+rayon = "1.10.0"
+rusqlite = { version = "0.37.0" }
+
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.140"
+
+tracing = { version = "0.1.41", features = ["attributes"] }
+tracing-subscriber = "0.3.19"
