+++ vadl-cosim/cosim-lib/src/ipc/cstructs.rs
