+++ vadl-cosim/cosim-lib/src/ipc/mod.rs
