+++ vadl-cosim/cosim-lib/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod cli;
+pub mod ipc;
+pub mod diff;
+pub mod cosim;
+pub mod trace;
+pub mod config;
