+++ vadl-cosim/cosim-lib/src/trace/db.rs
