+++ vadl-cosim/cosim-lib/src/trace/mod.rs
