+++ vadl-cosim/tui/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "tui"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
